depot_deploy: make child-state tracking reusable
