+++ repos/gems/src/app/depot_deploy/main.cc
@@ -19,78 +19,9 @@
 #include <os/reporter.h>
 
 /* local includes */
-#include "child.h"
+#include "children.h"
 
-namespace Depot_deploy {
-	struct Children;
-	struct Main;
-}
-
-
-class Depot_deploy::Children
-{
-	private:
-
-		Allocator &_alloc;
-
-		List_model<Child> _children { };
-
-		struct Model_update_policy : List_model<Child>::Update_policy
-		{
-			Allocator &_alloc;
-
-			Model_update_policy(Allocator &alloc) : _alloc(alloc) { }
-
-			void destroy_element(Child &c) { destroy(_alloc, &c); }
-
-			Child &create_element(Xml_node node)
-			{
-				return *new (_alloc) Child(_alloc, node);
-			}
-
-			void update_element(Child &c, Xml_node node) { c.apply_config(node); }
-
-			static bool element_matches_xml_node(Child const &child, Xml_node node)
-			{
-				return node.attribute_value("name", Child::Name()) == child.name();
-			}
-
-			static bool node_is_element(Xml_node node) { return node.has_type("start"); }
-
-		} _model_update_policy { _alloc };
-
-	public:
-
-		Children(Allocator &alloc) : _alloc(alloc) { }
-
-		void apply_config(Xml_node config)
-		{
-			_children.update_from_xml(_model_update_policy, config);
-		}
-
-		void apply_blueprint(Xml_node blueprint)
-		{
-			blueprint.for_each_sub_node("pkg", [&] (Xml_node pkg) {
-				_children.for_each([&] (Child &child) {
-					child.apply_blueprint(pkg); }); });
-
-			blueprint.for_each_sub_node("missing", [&] (Xml_node missing) {
-				_children.for_each([&] (Child &child) {
-					child.mark_as_incomplete(missing); }); });
-		}
-
-		void gen_start_nodes(Xml_generator &xml, Xml_node common)
-		{
-			_children.for_each([&] (Child const &child) {
-				child.gen_start_node(xml, common); });
-		}
-
-		void gen_queries(Xml_generator &xml)
-		{
-			_children.for_each([&] (Child const &child) {
-				child.gen_query(xml); });
-		}
-};
+namespace Depot_deploy { struct Main; }
 
 
 struct Depot_deploy::Main
@@ -102,9 +33,6 @@ struct Depot_deploy::Main
 
 	Expanding_reporter _query_reporter       { _env, "query" , "query"};
 	Expanding_reporter _init_config_reporter { _env, "config", "init.config"};
-
-	size_t _query_buffer_size       = 4096;
-	size_t _init_config_buffer_size = 4096;
 
 	Heap _heap { _env.ram(), _env.rm() };
 
@@ -135,7 +63,8 @@ struct Depot_deploy::Main
 		_init_config_reporter.generate([&] (Xml_generator &xml) {
 			Xml_node static_config = config.sub_node("static");
 			xml.append(static_config.content_base(), static_config.content_size());
-			_children.gen_start_nodes(xml, config.sub_node("common_routes"));
+			Child::Depot_rom_server const parent { };
+			_children.gen_start_nodes(xml, config.sub_node("common_routes"), parent);
 		});
 
 		/* update query for blueprints of all unconfigured start nodes */
